moved config to app in prep for separate filecreate(or similar) config
