+++ pkg/config/app/app_config_loader_test.go
@@ -1,4 +1,4 @@
-package config
+package app
 
 import (
 	"os"
