+++ pkg/config/app/user_settings.go
@@ -1,4 +1,4 @@
-package config
+package app
 
 // UserSettings contains user-specific settings
 type UserSettings struct {
