+++ pkg/config/app/validator.go
@@ -1,4 +1,4 @@
-package config
+package app
 
 import (
 	"errors"
