+++ pkg/config/app/validator_test.go
@@ -1,4 +1,4 @@
-package config
+package app
 
 import "testing"
