+++ pkg/fileops/fileops.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/matthewchivers/journal/pkg/config"
+	config "github.com/matthewchivers/journal/pkg/config/app"
 	"github.com/matthewchivers/journal/pkg/templating"
 )
