+++ pkg/fileops/fileops_test.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/matthewchivers/journal/pkg/config"
+	config "github.com/matthewchivers/journal/pkg/config/app"
 	"github.com/matthewchivers/journal/pkg/paths"
 	"github.com/stretchr/testify/assert"
 )
