+++ pkg/templating/pattern_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/matthewchivers/journal/pkg/caltools"
-	"github.com/matthewchivers/journal/pkg/config"
+	"github.com/matthewchivers/journal/pkg/config/app"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -25,56 +25,56 @@ func TestParsePattern(t *testing.T) {
 	tests := []struct {
 		name         string
 		pathTemplate string
-		fileType     config.FileType
+		fileType     app.FileType
 		want         string
 		wantErr      bool
 	}{
 		{
 			name:         "Valid template: year/month/day/",
 			pathTemplate: "{{.Year}}/{{.Month}}/{{.Day}}/",
-			fileType:     config.FileType{},
+			fileType:     app.FileType{},
 			want:         currentYear + "/" + currentMonth + "/" + currentDay + "/",
 			wantErr:      false,
 		},
 		{
 			name:         "Valid template: year/month/week/FileTypeName/",
 			pathTemplate: "{{.Year}}/{{.Month}}/{{.WeekNumber}}/{{.FileTypeName}}/",
-			fileType:     config.FileType{Name: "foo-note"},
+			fileType:     app.FileType{Name: "foo-note"},
 			want:         currentYear + "/" + currentMonth + "/" + string(rune(currentWeek)) + "/foo-note/",
 			wantErr:      false,
 		},
 		{
 			name:         "Valid template: year/month/wc weekCommencing/weekday/template/",
 			pathTemplate: "{{.Year}}/{{.Month}}/wc {{.WeekCommencing}}/{{.WeekdayName}}/{{.FileTypeName}}/",
-			fileType:     config.FileType{Name: "foo-note"},
+			fileType:     app.FileType{Name: "foo-note"},
 			want:         wcCurrentYear + "/" + wcCurrentMonth + "/wc " + weekCommencing.Format("2006-01-02") + "/" + currentWeekdayName + "/foo-note/",
 			wantErr:      false,
 		},
 		{
 			name:         "Valid template: year/month/week-commencing/template/",
 			pathTemplate: "{{.Year}}/{{.Month}}/{{.WeekCommencing}}/{{.FileTypeName}}/",
-			fileType:     config.FileType{Name: "foo-note"},
+			fileType:     app.FileType{Name: "foo-note"},
 			want:         currentYear + "/" + currentMonth + "/" + weekCommencing.Format("2006-01-02") + "/foo-note/",
 			wantErr:      false,
 		},
 		{
 			name:         "Path template without placeholders",
 			pathTemplate: "static-path/foo-note.md",
-			fileType:     config.FileType{Name: "foo-note"},
+			fileType:     app.FileType{Name: "foo-note"},
 			want:         "static-path/foo-note.md",
 			wantErr:      false,
 		},
 		{
 			name:         "Invalid path template",
 			pathTemplate: "{{.Year}/{{.Month}}/{{.Day}}/{{.FileTypeName}}/", // Missing a closing brace
-			fileType:     config.FileType{Name: "foo-note"},
+			fileType:     app.FileType{Name: "foo-note"},
 			want:         "",
 			wantErr:      true,
 		},
 		{
 			name:         "Empty path template",
 			pathTemplate: "",
-			fileType:     config.FileType{},
+			fileType:     app.FileType{},
 			want:         "",
 			wantErr:      false,
 		},
